[bug 1214396] Move sumo/api.py to sumo/api_utils.py
This will fix some issues with bootstrapping when we start adding api
view code to sumo/api.py.
