+++ kitsune/search/api.py
@@ -8,7 +8,7 @@ from kitsune.products.models import Product
 from kitsune.questions.models import Question, QuestionMappingType
 from kitsune.questions.api import QuestionSerializer
 from kitsune.search import es_utils
-from kitsune.sumo.api import GenericAPIException
+from kitsune.sumo.api_utils import GenericAPIException
 from kitsune.wiki.api import DocumentDetailSerializer
 from kitsune.wiki.models import Document, DocumentMappingType
