+++ kitsune/users/api.py
@@ -24,7 +24,7 @@ from kitsune.access.decorators import login_required
 from kitsune.questions.models import Answer
 from kitsune.questions.utils import num_answers, num_solutions, num_questions
 from kitsune.sumo import email_utils
-from kitsune.sumo.api import DateTimeUTCField, GenericAPIException, PermissionMod
+from kitsune.sumo.api_utils import DateTimeUTCField, GenericAPIException, PermissionMod
 from kitsune.sumo.decorators import json_view
 from kitsune.users.helpers import profile_avatar
 from kitsune.users.models import Profile, RegistrationProfile, Setting
